Merge from master

+++ wars/api-backoffice/src/main/java/it/govpay/backoffice/v1/beans/ListaIncassiIndex.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaIncassiIndex extends Lista<IncassoIndex> {
 	
-	public ListaIncassiIndex(List<IncassoIndex> incassi, URI requestUri, long count, long offset, long limit) {
+	public ListaIncassiIndex(List<IncassoIndex> incassi, URI requestUri, Long count, Integer offset, Integer limit) {
 		super(incassi, requestUri, count, offset, limit);
 	}
