+++ wars/api-pagamento/src/main/java/it/govpay/pagamento/v2/beans/ListaPendenzeIndex.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaPendenzeIndex extends Lista<PendenzaIndex> {
 	
-	public ListaPendenzeIndex(List<PendenzaIndex> pagamentiPortale, URI requestUri, long count, long pagina, long limit) {
+	public ListaPendenzeIndex(List<PendenzaIndex> pagamentiPortale, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(pagamentiPortale, requestUri, count, pagina, limit);
 	}
